requirements.txt
alembic==0.8.9
click==6.7
Flask
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-Script==2.0.5
Flask-Login==0.4.1
httplib2
itsdangerous==0.24
Jinja2
Mako==1.0.6
MarkupSafe==0.23
oauth2==1.9.0.post1
pycparser==2.14
python-editor==1.0.3
requests==2.20.0
webassets==0.12.0
Werkzeug
gunicorn==19.6.0
beautifulsoup4==4.5.3
line-bot-sdk
imgurpython==1.1.7
psycopg2
